cl_options: add -fdebug-prefix-map
Try to follow clang, accepting pairs of <old>=<new> paths.

dibuilder: use debugPrefixMap to remap source locs

This is required to have reproducible builds without remapping
the source code directory to a fixed location (for example in Bazel,
where the source code for every build action will end up in a
sanbox directory with some hash in the path name).

Try to keep the implementation as close as possible to what clang
does.