getPointerBitmap: change the result if there are no ptrs
WARN: this changes the semantics of __traits(getPointersBitmap)
so callers need to be aware of it.

getPointerBitmap builds an array literal for the bitmap, which is
proportional to the size of the type in question.

And there is no much sense to do this for types that are known to
contain no pointers anyway: we are just building a huge array literal
filled with zeroes.